refactor(arch): Phase 2 — PR #478 리뷰 3차 수정
- B1: @internal 태그 제거 (email-token api.ts 2개 함수)
- S1: shared/email/index.ts barrel 생성, 프로덕션 코드 barrel import 전환
- S2: entities/email-token에서 shared 타입(EmailDispatcher/EmailMessage) re-export 제거
- S3: SKILL_TYPES/SKILL_CATEGORIES satisfies 체크 추가
- S4: dispatcher.test.ts colocated 이동
