+++ src/shared/email/index.ts
@@ -1,0 +1,2 @@
+export { createEmailDispatcher } from './dispatcher';
+export type { EmailDispatcher, EmailMessage } from './types';
